refactor: extract attachment response builder to lib/attachment-response.ts
